Fix drag/drop bug that caused click event
A drag/drop operation was causing a click event after the drop. The mousedown
event and the mouseup event were happening on th same element, so the browser
fires a click event.

The only way to prevent this it seems is to temporarily disable click events
ont he whole document.  It's a hack, but at the moment, it seems like a bug in
browser logic that we'll have to deal with.